+++ examples/koa/server.js
@@ -1,10 +1,11 @@
-'use strict';
+import Koa from 'koa';
+import sockjs from 'sockjs';
+import http from 'node:http';
+import fs from 'node:fs';
+import path from 'node:path';
+import url from 'node:url';
 
-const Koa = require('koa');
-const sockjs = require('sockjs');
-const http = require('http');
-const fs = require('fs');
-const path = require('path');
+const __dirname = path.dirname(url.fileURLToPath(import.meta.url));
 
 // 1. Echo sockjs server
 const sockjs_opts = {
@@ -22,7 +23,7 @@ const app = new Koa();
 
 app.use(function (ctx, next) {
   return next().then(() => {
-    const filePath = __dirname + '/index.html';
+    const filePath = path.join(__dirname, 'index.html');
     ctx.response.type = path.extname(filePath);
     ctx.response.body = fs.createReadStream(filePath);
   });